call_summarizer/requirements.txt
fastapi
uvicorn
httpx
torch
whisper
langgraph==0.0.41
autogen @ git+https://github.com/microsoft/autogen.git
python-dotenv